+++ datadog_checks_base/datadog_checks/base/checks/base.py
@@ -35,6 +35,7 @@ from ..utils.http import RequestsWrapper
 from ..utils.limiter import Limiter
 from ..utils.metadata import MetadataManager
 from ..utils.secrets import SecretsSanitizer
+from ..utils.tagging import GENERIC_TAGS
 from ..utils.tls import TlsContextWrapper
 
 try:
@@ -188,6 +189,9 @@ class AgentCheck(object):
         self.instance = instance  # type: InstanceType
         self.instances = instances  # type: List[InstanceType]
         self.warnings = []  # type: List[str]
+        self.disable_generic_tags = (
+            is_affirmative(self.instance.get('disable_generic_tags', False)) if instance else False
+        )
 
         # `self.hostname` is deprecated, use `datadog_agent.get_hostname()` instead
         self.hostname = datadog_agent.get_hostname()  # type: str
@@ -1087,13 +1091,21 @@ class AgentCheck(object):
         for tag in tags:
             if tag is None:
                 continue
-
             try:
                 tag = to_native_string(tag)
             except UnicodeError:
                 self.log.warning('Encoding error with tag `%s` for metric `%s`, ignoring tag', tag, metric_name)
                 continue
-
-            normalized_tags.append(tag)
-
+            if self.disable_generic_tags:
+                normalized_tags.append(self.degeneralise_tag(tag))
+            else:
+                normalized_tags.append(tag)
         return normalized_tags
+
+    def degeneralise_tag(self, tag):
+        tag_name, value = tag.split(':')
+        if tag_name in GENERIC_TAGS:
+            new_name = '{}_{}'.format(self.name, tag_name)
+            return '{}:{}'.format(new_name, value)
+        else:
+            return tag