+++ datadog_checks_dev/datadog_checks/dev/tooling/utils.py
@@ -311,8 +311,12 @@ def get_test_directory(check_name):
     return os.path.join(get_root(), check_name, 'tests')
 
 
+def get_codeowners_file():
+    return os.path.join(get_root(), '.github', 'CODEOWNERS')
+
+
 def get_codeowners():
-    codeowners_file = os.path.join(get_root(), '.github', 'CODEOWNERS')
+    codeowners_file = get_codeowners_file()
     contents = read_file_lines(codeowners_file)
     return contents
 
@@ -551,6 +555,8 @@ def has_e2e(check):
 
 def has_process_signature(check):
     manifest_file = get_manifest_file(check)
+    if not file_exists(manifest_file):
+        return False
     try:
         with open(manifest_file) as f:
             manifest = json.loads(f.read())
@@ -578,6 +584,8 @@ def has_recommended_monitor(check):
 
 def _has_asset_in_manifest(check, asset):
     manifest_file = get_manifest_file(check)
+    if not file_exists(manifest_file):
+        return False
     try:
         with open(manifest_file) as f:
             manifest = json.loads(f.read())