+++ directory/datadog_checks/directory/config_models/instance.py
@@ -21,6 +21,7 @@ class InstanceConfig(BaseModel):
     directory: str
     dirs_patterns_full: Optional[bool]
     dirtagname: Optional[str]
+    disable_generic_tags: Optional[bool]
     empty_default_hostname: Optional[bool]
     exclude_dirs: Optional[Sequence[str]]
     filegauges: Optional[bool]