+++ haproxy/datadog_checks/haproxy/config_models/defaults.py
@@ -76,6 +76,10 @@ def instance_count_status_by_service(field, value):
     return True
 
 
+def instance_disable_generic_tags(field, value):
+    return False
+
+
 def instance_disable_legacy_service_tag(field, value):
     return False