+++ kafka_consumer/datadog_checks/kafka_consumer/config_models/defaults.py
@@ -24,6 +24,10 @@ def instance_consumer_groups(field, value):
     return get_default_field_value(field, value)
 
 
+def instance_disable_generic_tags(field, value):
+    return False
+
+
 def instance_empty_default_hostname(field, value):
     return False
 
@@ -86,10 +90,6 @@ def instance_ssl_certfile(field, value):
 
 def instance_ssl_check_hostname(field, value):
     return True
-
-
-def instance_ssl_context(field, value):
-    return get_default_field_value(field, value)
 
 
 def instance_ssl_crlfile(field, value):