+++ kafka_consumer/datadog_checks/kafka_consumer/config_models/instance.py
@@ -19,6 +19,7 @@ class InstanceConfig(BaseModel):
 
     broker_requests_batch_size: Optional[int]
     consumer_groups: Optional[Mapping[str, Any]]
+    disable_generic_tags: Optional[bool]
     empty_default_hostname: Optional[bool]
     kafka_client_api_version: Optional[str]
     kafka_connect_str: Union[str, Sequence[str]]
@@ -36,7 +37,6 @@ class InstanceConfig(BaseModel):
     ssl_cafile: Optional[str]
     ssl_certfile: Optional[str]
     ssl_check_hostname: Optional[bool]
-    ssl_context: Optional[str]
     ssl_crlfile: Optional[str]
     ssl_keyfile: Optional[str]
     ssl_password: Optional[str]