+++ nagios/datadog_checks/nagios/config_models/defaults.py
@@ -20,6 +20,10 @@ def instance_collect_service_performance_data(field, value):
     return False
 
 
+def instance_disable_generic_tags(field, value):
+    return False
+
+
 def instance_empty_default_hostname(field, value):
     return False