+++ openldap/datadog_checks/openldap/config_models/instance.py
@@ -18,6 +18,7 @@ class InstanceConfig(BaseModel):
         allow_mutation = False
 
     custom_queries: Optional[Sequence[Mapping[str, Any]]]
+    disable_generic_tags: Optional[bool]
     empty_default_hostname: Optional[bool]
     min_collection_interval: Optional[float]
     password: Optional[str]