+++ oracle/datadog_checks/oracle/config_models/instance.py
@@ -27,6 +27,7 @@ class InstanceConfig(BaseModel):
         allow_mutation = False
 
     custom_queries: Optional[Sequence[CustomQuery]]
+    disable_generic_tags: Optional[bool]
     empty_default_hostname: Optional[bool]
     jdbc_driver_path: Optional[str]
     min_collection_interval: Optional[float]