+++ varnish/datadog_checks/varnish/config_models/instance.py
@@ -19,6 +19,7 @@ class InstanceConfig(BaseModel):
 
     daemon_host: Optional[str]
     daemon_port: Optional[int]
+    disable_generic_tags: Optional[bool]
     empty_default_hostname: Optional[bool]
     metrics_filter: Optional[Sequence[str]]
     min_collection_interval: Optional[float]